requirements.txt
# Web Scraping básico
requests>=2.31.0
beautifulsoup4>=4.12.2
lxml>=4.9.3

# Manejo de URLs y archivos
urllib3>=2.0.4

# Extracción de texto de PDFs
pdfminer.six>=20221105
PyMuPDF>=1.23.8
pdfplumber>=0.10.3

# OCR para PDFs escaneados
pytesseract>=0.3.10
Pillow>=10.0.1
pdf2image>=1.16.3

# Base de datos
pymongo>=4.5.0

# Utilidades
python-dateutil>=2.8.2
tqdm>=4.66.1

# Configuración
python-decouple>=3.8

# Logging mejorado
colorlog>=6.7.0